src/lambda-inference/requirements.txt
# List all python libraries for the lambda
tensorflow==2.1.3
tensorflow_hub
boto3
Pillow